Add scala 2.11 profile
This is only a profile for legacy projects and there is no guarantee
that everything will work properly when compiling with this profile - it
just configures all scala versions down to 2.11. It is only there to be
able to build a scala 2.11 variant of the lib on the local machine - not
for any automated publishing / pipeline workflow.
Call it with mvn -P scala-2.11 package and the corresponding jar in
target/ will be compiled against 2.11